[refactor][common] Move DingoSafeMap to common.
